python/langchain-weather/pyproject.toml
[project]
name = "langchain-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
]
